Move custom resources into own package
