+++ provider/pkg/resources/customresources/custom_keyvault_accesspolicy_test.go
@@ -1,4 +1,4 @@
-package resources
+package customresources
 
 import (
 	"testing"
