+++ provider/pkg/resources/customresources/custom_webapp_delete.go
@@ -1,4 +1,4 @@
-package resources
+package customresources
 
 import (
 	"context"
